+++ examples/data-sources/sra_security_provider/data-source.tf
@@ -1,0 +1,3 @@
+data "sra_security_provider" "supplier" {
+  name = "Supplier"
+}